refactor: move styled components into own file
